gpu-simulator/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
numpy>=1.21.0
pydantic>=2.0.0
pytest>=7.0.0
hypothesis>=6.80.0
python-multipart>=0.0.5
setuptools>=65.0.0
reportlab>=4.0.0
Pillow>=10.0.0
psutil>=5.9.0